backend/pyproject.toml
[tool.poetry.scripts]
start = "backend.main:main"

[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["vitor <vitor.azevedo009@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pylint = "^2.14.5"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
